working editor

+++ Seminarfacharbeit/src/ui/utils/MultipleInputDialog.java
@@ -56,7 +56,7 @@ public class MultipleInputDialog extends Dialog<Float[]>
 		
 		for (int i = 0; i < labels.length; i++)
 		{
-			NumberInputField field = new NumberInputField(Float.MAX_VALUE, Float.MIN_VALUE, false);
+			NumberInputField field = new NumberInputField(Integer.MAX_VALUE, Integer.MIN_VALUE, false);
 			numberFields[i] = field;
 			grid.add(labels[i], 0, i);
 			grid.add(field, 1, i);
